# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Olive v1 Core API
  version: 0.1.0
  description: |
    Canada-first dental AI scribe. Backend (`apps/api`) is source of truth.
    Clinic PHIPA fields are nullable and do not block local scaffold.
    Quebec Law 25 is out of v1.
  contact:
    name: Olive backend
servers:
  - url: http://localhost:3000
    description: Local compose + apps/api
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Demo clinic login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: od@demo.olive.local }
                password: { type: string, example: demo }
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema:
                type: object
                required: [token, expiresAt, user, clinic, flags]
                properties:
                  token: { type: string }
                  expiresAt: { type: string, format: date-time }
                  user: { $ref: "#/components/schemas/User" }
                  clinic: { $ref: "#/components/schemas/Clinic" }
                  flags: { $ref: "#/components/schemas/FeatureFlags" }
        "401":
          $ref: "#/components/responses/Error"
  /v1/session:
    get:
      tags: [Auth, Flags]
      summary: Session bootstrap (user + clinic + flags)
      responses:
        "200":
          description: Bootstrap payload
          content:
            application/json:
              schema:
                type: object
                required: [user, clinic, flags]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  clinic: { $ref: "#/components/schemas/Clinic" }
                  flags: { $ref: "#/components/schemas/FeatureFlags" }
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/clinic/audio/delete:
    post:
      tags: [Recording]
      summary: SEC-007 clinic / end-of-contract audio delete (admin). No notes/transcripts cascade. No auto-TTL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm: { type: string, example: delete-clinic-audio }
      responses:
        "200":
          description: Clinic audio deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }
                  notesUntouched: { type: boolean }
                  transcriptsUntouched: { type: boolean }
        "403":
          description: admin_required
...
  /v1/patients/{id}:
    get:
      tags: [Patients]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient: { $ref: "#/components/schemas/PatientProfile" }
        "404":
          $ref: "#/components/responses/Error"
  /v1/patients/{id}/profile:
    get:
      tags: [Patients]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile: { $ref: "#/components/schemas/PatientProfile" }
  /v1/patients/{id}/chat:
    get:
      tags: [Chat]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Secure thread
          content:
            application/json:
              schema:
                type: object
                required: [thread, messages]
                properties:
                  thread: { $ref: "#/components/schemas/ChatThread" }
                  messages:
                    type: array
                    items: { $ref: "#/components/schemas/ChatMessage" }
    post:
      tags: [Chat]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                visitId: { type: string, format: uuid }
      responses:
        "200":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /v1/patients/{id}/consents:
    post:
      tags: [Recording]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentInput"
      responses:
        "200":
          description: Consent stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consent"
        "422":
          $ref: "#/components/responses/Error"
  /v1/visits:
    post:
      tags: [Visits]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientId]
              properties:
                patientId: { type: string, format: uuid }
      responses:
        "200":
          description: Visit created (in_progress)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /v1/visits/{id}:
    get:
      tags: [Visits]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /v1/visits/{id}/end:
    post:
      tags: [Visits]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Visit completed. Audio is kept (clinic-controlled; no auto-TTL).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /v1/visits/{id}/consents:
    get:
      tags: [Recording]
      summary: OLI-6 consent evidence for the visit
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Consent evidence rows
          content:
            application/json:
              schema:
                type: object
                required: [consents]
                properties:
                  consents:
                    type: array
                    items: { $ref: "#/components/schemas/Consent" }
  /v1/visits/{id}/recording-gate:
    get:
      tags: [Recording]
      summary: Visit-scoped audio_capture gate
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Gate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingGate"
              examples:
                denied:
                  value: { allowed: false, reason: missing_audio_capture_consent }
                allowed:
                  value: { allowed: true, disclosureScriptId: audio-disclosure-v1 }
  /v1/visits/{id}/consent:
    post:
      tags: [Recording]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitConsentInput"
      responses:
        "200":
          description: Consent stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consent"
  /v1/visits/{id}/audio:
    post:
      tags: [Recording]
      summary: Ingest audio (refuses if recording-gate denied)
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bytesBase64: { type: string }
          audio/webm:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Asset + queued transcription job
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset: { $ref: "#/components/schemas/AudioAsset" }
                  job: { $ref: "#/components/schemas/TranscriptionJob" }
        "403":
          description: recording_gate_denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Recording]
      summary: Clinic-initiated visit audio delete (no auto-TTL). Does not cascade notes/transcripts.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }
                  notesPreserved: { type: integer }
                  transcriptsPreserved: { type: integer }
  /v1/visits/{id}/transcript:
    get:
      tags: [Transcript]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Transcript + segments + job
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcript: { $ref: "#/components/schemas/Transcript" }
                  segments:
                    type: array
                    items: { $ref: "#/components/schemas/TranscriptSegment" }
                  job: { $ref: "#/components/schemas/TranscriptionJob" }
  /v1/visits/{id}/transcript/stream:
    get:
      tags: [Transcript]
      summary: SSE segments with cursor replay
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: cursor
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
          description: Replay segments with seq > cursor
      responses:
        "200":
          description: |
            `text/event-stream`. Events:
            - `segment` (id = seq) `{ id, seq, speakerLabel, text, startMs, endMs }`
            - `status` `{ jobStatus, cursor }`
            - `done` `{ ok: true }` when job completed
          content:
            text/event-stream:
              schema:
                type: string
  /v1/visits/{id}/note:
    get:
      tags: [Notes]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Draft or signed note (created on first get)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      tags: [Notes]
      summary: Draft only — 403 if signed
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "403":
          description: note_signed_immutable
  /v1/visits/{id}/note/sign:
    post:
      tags: [Notes]
      summary: Explicit sign — never auto-sign
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Immutable snapshot frozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /v1/visits/{id}/follow-ups:
    get:
      tags: [FollowUps]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Follow-ups for the visit
          content:
            application/json:
              schema:
                type: object
                required: [followUps]
                properties:
                  followUps:
                    type: array
                    items: { $ref: "#/components/schemas/FollowUp" }
    post:
      tags: [FollowUps]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                messageClass:
                  $ref: "#/components/schemas/MessageClass"
      responses:
        "200":
          description: Draft follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /v1/follow-ups/{id}:
    patch:
      tags: [FollowUps]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /v1/follow-ups/{id}/skip:
    post:
      tags: [FollowUps]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Skipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /v1/follow-ups/{id}/edits:
    post:
      tags: [FollowUps]
      summary: Tone/edit event (no ML in v1)
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [before, after]
              properties:
                before: { type: string }
                after: { type: string }
      responses:
        "200":
          description: Edit event stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUpEdit"
  /v1/follow-ups/{id}/send:
    post:
      tags: [FollowUps]
      summary: CASL send gate (fake MessagingVendor — logs/queues, no real SMS)
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Sent via fake vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
        "403":
          description: |
            Fail-closed codes: `missing_messaging_consent`, `promotional_fail_closed`,
            `stop_fail_closed`, `unsigned_note`, `missing_clinic_identity`
  /v1/days/finish:
    post:
      tags: [Days]
      summary: Snapshot pending follow-up queue (FE is not SoT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date]
              properties:
                date: { type: string, format: date, example: "2026-09-19" }
      responses:
        "200":
          description: Day close
          content:
            application/json:
              schema:
                type: object
                properties:
                  dayClose: { type: object }
                  pendingCount: { type: integer }
                  snapshot: { type: object }
...
  /v1/pms/import:
    post:
      tags: [PMS]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                source: { type: string }
                patients:
                  type: array
                  items: { $ref: "#/components/schemas/ImportedPatient" }
      responses:
        "200":
          description: Import job
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: integer }
                  updated: { type: integer }
                  job: { type: object }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [dentist, staff, admin] }
        clinicId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid, description: "Same as clinicId (OLI-12). MFA is Wave B." }
    Clinic:
      type: object
      description: PHIPA agent/ESP fields are nullable and non-blocking for local scaffold
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid, description: Same as id }
        name: { type: string }
        legalName: { type: string }
        smsIdentity: { type: string }
        phone: { type: string }
        residencyRegion: { type: string, example: ca-central-1 }
        country: { type: string, example: CA }
        province: { type: string, example: ON }
        phipaAgreementVersion: { type: string, nullable: true }
        phipaAgreementAckedAt: { type: string, format: date-time, nullable: true }
        phipaAgreementAckedBy: { type: string, format: uuid, nullable: true }
...
    MessageClass:
      type: string
      enum: [clinical_transactional, promotional]
...